apps/mesh-daemon/pyproject.toml
[tool.poetry]
name = "offsec-mesh-daemon"
version = "0.1.0"
description = "OffSec Shield mesh federation daemon (root + proof gossiper)"
authors = ["Sovereign <you@example.com>"]
readme = "README.md"
packages = [{ include = "mesh_daemon" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.0"
blake3 = "^0.4.1"
PyNaCl = "^1.5.0"

[tool.poetry.scripts]
mesh-daemon = "mesh_daemon.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
